// frontend/src/app/models/report.ts
export default class Report {
    id: number;
    doctor: string;
    patient: string;
    date: string;
    time: string;
    doctorName: string;
    doctorLastname: string;
    doctorSpecialty: string;
    symptoms: string;
    diagnosis: string;
    therapy: string;
    nextExamination: string;
    examId: number;
}